0030655: Modeling Data - Provide interfaces for selection of the elements from BVH tree
Provide the easy to use interfaces for selection of the elements from BVH tree.
The selection rules should be implemented in the selector class derived from *BVH_BoxSet<>::Selector* or in *BVH_BoxSet<>::PairSelector* in Reject/Accept methods.

The *BVH_BoxSet<>::Selector* is used for selection of the elements from the tree.
The *BVH_BoxSet<>::PairSelector* is used for selection of the pairs of elements from two BVH trees.

Auxiliary changes:
- Two methods BVH_Box::IsOut(OtherBox) and BVH_Box::IsOut(Point) have been added;
- Added methods for conversion of Bnd boxes to BVH boxes

Test case for the issue.

(cherry picked from commit 9175060417729936736392d52444e8f751bcf986)

# Conflicts:
#	src/BVH/BVH_Box.hxx
#	tests/lowalgos/grids.list